pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "eth-account>=0.13.7",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "uvicorn>=0.35.0",
]
